+++ qtNanomite/qtDLGWindowView.cpp
@@ -43,7 +43,7 @@ qtDLGWindowView::qtDLGWindowView(QWidget *parent, Qt::WFlags flags,qint32 proces
 	m_pMainWindow = qtDLGNanomite::GetInstance();
 	m_processCountEnd = m_pMainWindow->coreDebugger->PIDs.size();
 
-	for(size_t i = 0; i < m_pMainWindow->coreDebugger->PIDs.size(); i++)
+	for(int i = 0; i < m_pMainWindow->coreDebugger->PIDs.size(); i++)
 	{
 		if(m_pMainWindow->coreDebugger->PIDs[i].dwPID == m_processID)
 		{
@@ -68,7 +68,7 @@ void qtDLGWindowView::EnumWindow()
 {
 	tblWindowView->setRowCount(0);
 
-	for(size_t i = m_processCountEntry; i < m_processCountEnd;i++)
+	for(int i = m_processCountEntry; i < m_processCountEnd;i++)
 	{
 		EnumWindows((WNDENUMPROC)EnumWindowCallBack,(LPARAM)m_pMainWindow->coreDebugger->PIDs[i].dwPID);
 	}
